requirements.txt
numpy
opencv-python
pillow
scikit-learn